IR_Project/requirements.txt
numpy>=1.24.0
faiss-cpu>=1.7.4
langchain>=0.1.0
openai>=1.12.0
pdfplumber>=0.10.3
pandas>=2.0.0